Refactor DSL v4 into focused modules
